Move LegacyControlUnit entity to control_unit domain in Backend
