+++ backend/src/main/kotlin/fr/gouv/cnsp/monitorfish/domain/entities/control_unit/LegacyControlUnit.kt
@@ -1,5 +1,6 @@
-package fr.gouv.cnsp.monitorfish.domain.entities.mission
+package fr.gouv.cnsp.monitorfish.domain.entities.control_units
 
+import fr.gouv.cnsp.monitorfish.domain.entities.mission.ControlResource
 import kotlinx.serialization.Serializable
 
 @Serializable
