+++ backend/src/main/kotlin/fr/gouv/cnsp/monitorfish/domain/entities/reporting/ReportingAndOccurrences.kt
@@ -1,6 +1,6 @@
 package fr.gouv.cnsp.monitorfish.domain.entities.reporting
 
-import fr.gouv.cnsp.monitorfish.domain.entities.mission.LegacyControlUnit
+import fr.gouv.cnsp.monitorfish.domain.entities.control_units.LegacyControlUnit
 
 data class ReportingAndOccurrences(
     val otherOccurrencesOfSameAlert: List<Reporting>,
