+++ backend/src/main/kotlin/fr/gouv/cnsp/monitorfish/domain/repositories/LegacyControlUnitRepository.kt
@@ -1,6 +1,6 @@
 package fr.gouv.cnsp.monitorfish.domain.repositories
 
-import fr.gouv.cnsp.monitorfish.domain.entities.mission.LegacyControlUnit
+import fr.gouv.cnsp.monitorfish.domain.entities.control_units.LegacyControlUnit
 
 interface LegacyControlUnitRepository {
     fun findAll(): List<LegacyControlUnit>
